+++ src/client.ts
@@ -3147,10 +3147,11 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     return this.get<{ campaign: CampaignResponse } & APIResponse>(this.baseURL + `/campaigns/${id}`);
   }
 
-  async startCampaign(id: string, scheduledFor?: string) {
+  async startCampaign(id: string, options?: { scheduledFor?: string; stopAt?: string }) {
     this.validateServerSideAuth();
     return this.post<{ campaign: CampaignResponse } & APIResponse>(this.baseURL + `/campaigns/${id}/start`, {
-      scheduled_for: scheduledFor,
+      scheduled_for: options?.scheduledFor,
+      stop_at: options?.stopAt,
     });
   }
   /**
@@ -3161,12 +3162,12 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    */
   async queryCampaigns(filter: CampaignFilters, sort?: CampaignSort, options?: CampaignQueryOptions) {
     this.validateServerSideAuth();
-    return await this.post<{
-      campaigns: CampaignResponse[];
-      segments: Record<string, Segment>;
-      channels?: Record<string, ChannelResponse<StreamChatGenerics>>;
-      users?: Record<string, UserResponse<StreamChatGenerics>>;
-    }>(this.baseURL + `/campaigns/query`, {
+    return await this.post<
+      {
+        campaigns: CampaignResponse[];
+        next?: string;
+      } & APIResponse
+    >(this.baseURL + `/campaigns/query`, {
       filter,
       sort,
       ...(options || {}),